# swagger-ui/swagger-api.yaml
swagger: '2.0'
info:
  version: 0.0.1
  title: Engine Hub Service
  description: An API specification of Engine Hub Service based on swagger-2.0 specification
  contact:
    name: Aid
    email: aid666@gmail.com
  license:
    name: MIT
    url: 'http://opensource.org/licenses/MIT'
basePath: /hub
schemes:
  - http
consumes:
  - application/json
  - application/xml
produces:
  - application/json
  - application/xml
paths:

  /algorithms:
    get:
      tags:
      - Algorithms
      summary: Search processers metdata.
      description: Processers, it is actually execute code of any process flow. But this processer do not keep any model or paramter.
      parameters:
      - in: query
        name: catalog
        type: string
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/ProcesserAlgorithm'

  /adhoc:
    post:
      tags:
      - Ad_hoc
      parameters:
      - name: body
        in: body
        schema:
          $ref: '#/definitions/ProcessAdhocExe'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/ProcessMetadata'

  /adhoc/{exeKey}:
    get:
      tags:
      - Ad_hoc
      parameters:
      - name: exeKey
        in: path
        type: string
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/ProcessMetadata'

  /flow:
    get:
      tags:
      - Flow
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/ProcessFlowMetadata'
    post:
      tags:
      - Flow
      parameters:
      - name: body
        in: body
        schema:
          $ref: '#/definitions/ProcessFlow'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/ProcessFlowMetadata'

  /flow/{flowKey}:
    get:
      tags:
      - Flow
      parameters:
      - name: flowKey
        in: path
        type: string
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/ProcessFlowMetadata'

  /flow/{flowKey}/processes:
    get:
      tags:
      - Flow
      - Processes
      parameters:
      - name: flowKey
        in: path
        type: string
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/ProcessMetadata'
    post:
      tags:
      - Flow
      - Processes
      parameters:
      - name: flowKey
        in: path
        type: string
      - name: body
        in: body
        schema:
          $ref: '#/definitions/DMData'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/ProcessMetadata'


  /flow/{flowKey}/processes/{procKey}:
    get:
      tags:
      - Flow
      - Processes
      parameters:
      - name: flowKey
        in: path
        type: string
      - name: procKey
        in: path
        type: string
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/ProcessMetadata'
    delete:
      tags:
      - Flow
      - Processes
      parameters:
      - name: flowKey
        in: path
        type: string
      - name: procKey
        in: path
        type: string
      responses:
        204:
          description: OK
  /queue:
    get:
      tags:
      - queue
      summary: Query the queue of process which waiting to be running.
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/WaitingProcessMetadata'

  /processes/{pKey}/flow:
    get:
      tags:
      - Processes
      summary: Get the flow of the process
      parameters:
      - name: pKey
        in: path
        type: string
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/ProcessFlow'

  /processes/{pKey}/data:
    get:
      tags:
      - Processes
      summary: Get the data of the process
      parameters:
      - name: pKey
        in: path
        type: string
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/DMData'

  /processes/{pKey}/checklogs:
    get:
      tags:
      - Processes
      summary: Check the check log of a process
      parameters:
      - name: pKey
        in: path
        type: string
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/WaitingProcessCheckLog'
    post:
      tags:
      - Processes
      summary: Add a check log of a process
      parameters:
      - name: pKey
        in: path
        type: string
      - name: body
        in: body
        schema:
          $ref: '#/definitions/WaitingProcessCheckLog'
      responses:
        204:
          description: OK
        404:
          description: No such WP in the queue, this could be an error, or the queue had been marked as running.


  /processes/{processKey}:
    get:
      tags:
      - Processes
      parameters:
      - name: processKey
        in: path
        type: string
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/ProcessStatus'

  /processes/{processKey}/result:
    get:
      tags:
      - Processes
      parameters:
        - name: processKey
          in: path
          type: string
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/Process'

definitions:
  Domain:
    properties:
      name:
        type: string
      date:
        description: The domain in differnent date may different, think about the change of computer brand from 1995 to 2015 .
        type: string
      parentDomain:
        $ref: '#/definitions/Domain'

  DMData:
    properties:
      key:
        type: string
      data:
        type: array
        items:
         $ref: '#/definitions/ColumnData'

  ColumnData:
    properties:
      attribute:
        $ref: '#/definitions/Attribute'
      data:
        type: array
        items:
          type: string

  DataInfor:
    properties:
      name:
        type: string
      domain:
        $ref: '#/definitions/Domain'

  Attribute:
    properties:
      name:
        type: string
      type:
        type: string
      format:
        type: string
      dataInfor:
        $ref: '#/definitions/DataInfor'

  ProcessMetadata:
    properties:
      key:
        type: string
      status:
        type: string

  ProcessStatus:
    properties:
      key:
        type: string
      status:
        type: string

  Process:
    properties:
      key:
        type: string
      inputData:
        $ref: '#/definitions/DMData'
      outputData:
        $ref: '#/definitions/DMData'
      errors:
        type: array
        items:
          type: string
      warnings:
        type: array
        items:
          type: string
      logs:
        type: array
        items:
          type: string

  ProcesserAlgorithm:
    properties:
      key:
        type: string
      name:
        type: string
      description:
        type: string
      tags:
        type: array
        items:
          type: string
      catalog:
        description: The catelog path of the trainer, classifier.decisionTree.id3
        type: string

  ProcessAdhocExe:
    properties:
      flow:
        $ref: '#/definitions/ProcessFlow'
      data:
        $ref: '#/definitions/DMData'

  ProcessFlow:
    properties:
      key:
        type: string
      processers:
        type: array
        items:
          $ref: '#/definitions/Processer'

  Processer:
    properties:
      algKey:
        type: string
      model:
        $ref: '#/definitions/DMData'

  ProcessFlowMetadata:
    properties:
      key:
        type: string

  WaitingProcessMetadata:
    properties:
      key:
        type: string
      algorithms:
        type: array
        items:
          type: string

  WaitingProcessCheckLog:
    properties:
      datetime:
        type: string
      status:
        description: DENY, ACCEPT
        type: string
      description:
        type: string
      checkerKey:
        type: string